Initial commit + a beginning
My first commit for my portfolio website that I will be using. I've already started with making the projects.html page where I will be displaying all my work, a dark mode and light mode has been made and is working. Has still work to do but for now this is my start. [This repository will be writen in english purely because of the website being strict english].