Projects page being filled & private page made
I began filling up the projects page with some of my work that i've made. This is still in the working and has to be finished, also made a private.html page because of one of the projects the github repository being private explaining why there isn't a public repository to be viewed.