Scrapped now projects only
I've decided to scrap this project and make it into a little projects place where I can show my work, my reasoning for this is because I decided to make a different React website for my portfolio, still this looks good in my opinion so will be considered work. End.